Move styles around
