+++ src/client/components/index.tsx
@@ -1,2 +1,3 @@
 export * from "./SimpleDatepicker";
 export * from "./ContainerWrapper";
+export * from "./Loader";
