+++ src/client/logo_white.png
