+++ src/client/navbar/Navbar.less
@@ -33,7 +33,7 @@ nav.on-board-nav {
             .logo {
                 width: 100px;
                 height: 50px;
-                background-image: url("../styles/logo_white.png");
+                background-image: url("../logo_white.png");
                 background-position: center;
                 background-repeat: no-repeat;
                 background-size: contain;
