+++ src/client/user/Login.less
@@ -175,7 +175,7 @@
     }
 
     .logo-image {
-        background-image: url("logo_white.png");
+        background-image: url("../logo_white.png");
         background-position: center;
         background-repeat: no-repeat;
         background-size: contain;
